move to gss.typing
