+++ cirq-superstaq/cirq_superstaq/service.py
@@ -412,7 +412,7 @@ class Service(gss.service.Service):
         supports_compile: Optional[bool] = None,
         available: Optional[bool] = None,
         retired: Optional[bool] = None,
-    ) -> List[gss.superstaq_client.TargetInfo]:
+    ) -> List[gss.typing.TargetInfo]:
         """Gets a list of Superstaq targets along with their status information.
 
         Args:
