+++ general-superstaq/general_superstaq/testing.py
@@ -1,4 +1,4 @@
-from general_superstaq.superstaq_client import TargetInfo
+from general_superstaq.typing import TargetInfo
 
 TARGET_LIST = {
     "aqt_keysight_qpu": {
